Move StoreBuilders to server/builders.go
